+++ src/date/advance.js
@@ -1,5 +1,5 @@
 import { isNumber, isString } from '../util/typeChecks';
-import { assertDate } from '../util/assertions';
+import { assertValidDate } from '../util/assertions';
 import { advanceDate } from './util/shift';
 import { normalizeProps, getPropsFromString } from './util/props';
 
@@ -48,7 +48,7 @@ import { normalizeProps, getPropsFromString } from './util/props';
  *
  **/
 export default function advance(date, props = {}, reset = false) {
-  assertDate(date);
+  assertValidDate(date);
   if (isNumber(props)) {
     date.setTime(date.getTime() + props);
   } else {