+++ src/date/define.js
@@ -12,6 +12,7 @@ defineInstance('isPast', methods.isPast);
 defineInstance('isValid', methods.isValid);
 defineInstance('isWeekday', methods.isWeekday);
 defineInstance('isWeekend', methods.isWeekend);
+defineInstance('relative', methods.relative);
 defineInstance('rewind', methods.rewind);
 defineInstance('set', methods.set);
 defineInstance('setDay', methods.setWeekday);