+++ src/date/methods.js
@@ -19,6 +19,7 @@ export { default as isPast } from './isPast';
 export { default as isValid } from './isValid';
 export { default as isWeekday } from './isWeekday';
 export { default as isWeekend } from './isWeekend';
+export { default as relative } from './relative';
 export { default as rewind } from './rewind';
 export { default as set } from './set';
 export { default as setISOWeek } from './setISOWeek';