+++ src/date/util/reset.js
@@ -7,10 +7,15 @@ export function resetByUnit(date, unit, end) {
 
 export function resetBySpecificity(date, specificity, end = false) {
   for (let i = specificity + 1; i < SPECIFICITY_INDEX.length; i++) {
-    const unit = SPECIFICITY_INDEX[i];
-    if (unit !== 'week') {
-      const val = getUnitEdge(unit, end, date);
-      callDateSet(date, unit, val);
+    let unit = SPECIFICITY_INDEX[i];
+    // When resetting by specificity always ignore weeks and reset
+    // the calendar date instead.
+    if (unit === 'week') {
+      continue;
+    } else if (unit === 'day') {
+      unit = 'date';
     }
+    const val = getUnitEdge(unit, end, date);
+    callDateSet(date, unit, val);
   }
 }